+++ src/ids.rs
@@ -1,0 +1,31 @@
+/*
+amFOSS Daemon: A discord bot for the amFOSS Discord server.
+Copyright (C) 2024 amFOSS
+
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program.  If not, see <https://www.gnu.org/licenses/>.
+*/
+pub const ARCHIVE_MESSAGE_ID: u64 = 1298636092886749294;
+pub const ARCHIVE_ROLE_ID: u64 = 1208457364274028574;
+pub const MOBILE_ROLE_ID: u64 = 1298553701094395936;
+pub const SYSTEMS_ROLE_ID: u64 = 1298553801191718944;
+pub const AI_ROLE_ID: u64 = 1298553753523453952;
+pub const RESEARCH_ROLE_ID: u64 = 1298553855474270219;
+pub const DEVOPS_ROLE_ID: u64 = 1298553883169132554;
+pub const WEB_ROLE_ID: u64 = 1298553910167994428;
+
+pub const GROUP_ONE_CHANNEL_ID: u64 = 1225098248293716008;
+pub const GROUP_TWO_CHANNEL_ID: u64 = 1225098298935738489;
+pub const GROUP_THREE_CHANNEL_ID: u64 = 1225098353378070710;
+pub const GROUP_FOUR_CHANNEL_ID: u64 = 1225098407216156712;
+pub const STATUS_UPDATE_CHANNEL_ID: u64 = 764575524127244318;
