+++ src/main.rs
@@ -16,39 +16,50 @@ You should have received a copy of the GNU General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
 mod commands;
-mod graphql;
+mod ids;
 mod scheduler;
-mod tasks;
 mod utils;
 
+use crate::ids::{
+    AI_ROLE_ID, ARCHIVE_MESSAGE_ID, ARCHIVE_ROLE_ID, DEVOPS_ROLE_ID, MOBILE_ROLE_ID,
+    RESEARCH_ROLE_ID, SYSTEMS_ROLE_ID, WEB_ROLE_ID,
+};
 use anyhow::Context as _;
-use std::collections::HashMap;
-
 use poise::{Context as PoiseContext, Framework, FrameworkOptions, PrefixFrameworkOptions};
 use serenity::{
-        client::{Context as SerenityContext, FullEvent}, model::{
-        channel::ReactionType,
-        gateway::GatewayIntents,
-        id::{MessageId, RoleId},
-    }
+    all::{ReactionType, RoleId},
+    client::{Context as SerenityContext, FullEvent},
+    model::{gateway::GatewayIntents, id::MessageId},
 };
+use std::collections::HashMap;
 
-type Context<'a> = PoiseContext<'a, Data, Error>;
-type Error = Box<dyn std::error::Error + Send + Sync>;
+pub type Error = Box<dyn std::error::Error + Send + Sync>;
+pub type Context<'a> = PoiseContext<'a, Data, Error>;
 
-struct Data {
-    reaction_roles: HashMap<ReactionType, RoleId>,
+pub struct Data {
+    pub reaction_roles: HashMap<ReactionType, RoleId>,
 }
 
-const ARCHIVE_MESSAGE_ID: u64 = 1298636092886749294;
-const ARCHIVE_ROLE_ID: u64 = 1208457364274028574;
-const MOBILE_ROLE_ID: u64 = 1298553701094395936;
-const SYSTEMS_ROLE_ID: u64 = 1298553801191718944;
-const AI_ROLE_ID: u64 = 1298553753523453952;
-const RESEARCH_ROLE_ID: u64 = 1298553855474270219;
-const DEVOPS_ROLE_ID: u64 = 1298553883169132554;
-const WEB_ROLE_ID: u64 = 1298553910167994428;
+pub fn initialize_data() -> Data {
+    let mut data = Data {
+        reaction_roles: HashMap::new(),
+    };
 
+   let roles = [
+        (ReactionType::Unicode("📁".to_string()), RoleId::new(ARCHIVE_ROLE_ID)),
+        (ReactionType::Unicode("📱".to_string()), RoleId::new(MOBILE_ROLE_ID)),
+        (ReactionType::Unicode("⚙️".to_string()), RoleId::new(SYSTEMS_ROLE_ID)),
+        (ReactionType::Unicode("🤖".to_string()), RoleId::new(AI_ROLE_ID)),
+        (ReactionType::Unicode("📜".to_string()), RoleId::new(RESEARCH_ROLE_ID)),
+        (ReactionType::Unicode("🚀".to_string()), RoleId::new(DEVOPS_ROLE_ID)),
+        (ReactionType::Unicode("🌐".to_string()), RoleId::new(WEB_ROLE_ID)),
+    ];
+
+    data.reaction_roles
+        .extend::<HashMap<ReactionType, RoleId>>(roles.into());
+
+    data
+}
 #[shuttle_runtime::main]
 async fn main(
     #[shuttle_runtime::Secrets] secret_store: shuttle_runtime::SecretStore,
@@ -74,7 +85,7 @@ async fn main(
                 poise::builtins::register_globally(ctx, &framework.options().commands).await?;
 
                 let data = initialize_data();
-                crate::scheduler::run_scheduler(ctx.clone()).await;
+                scheduler::run_scheduler(ctx.clone()).await;
 
                 Ok(data)
             })
@@ -92,35 +103,6 @@ async fn main(
     Ok(client.into())
 }
 
-fn initialize_data() -> Data {
-    let mut data = Data {
-        reaction_roles: HashMap::new(),
-    };
-
-    let archive_role_id = RoleId::new(ARCHIVE_ROLE_ID);
-    let mobile_role_id = RoleId::new(MOBILE_ROLE_ID);
-    let systems_role_id = RoleId::new(SYSTEMS_ROLE_ID);
-    let ai_role_id = RoleId::new(AI_ROLE_ID);
-    let research_role_id = RoleId::new(RESEARCH_ROLE_ID);
-    let devops_role_id = RoleId::new(DEVOPS_ROLE_ID);
-    let web_role_id = RoleId::new(WEB_ROLE_ID);
-
-
-    let message_roles = [
-        (ReactionType::Unicode("📁".to_string()), archive_role_id),
-        (ReactionType::Unicode("📱".to_string()), mobile_role_id),
-        (ReactionType::Unicode("⚙️".to_string()), systems_role_id),
-        (ReactionType::Unicode("🤖".to_string()), ai_role_id),
-        (ReactionType::Unicode("📜".to_string()), research_role_id),
-        (ReactionType::Unicode("🚀".to_string()), devops_role_id),
-        (ReactionType::Unicode("🌐".to_string()), web_role_id),
-   ];
-
-    data.reaction_roles.extend::<HashMap<ReactionType, RoleId>>(message_roles.into());
-
-    data
-}
-
 async fn event_handler(
     ctx: &SerenityContext,
     event: &FullEvent,
@@ -130,31 +112,50 @@ async fn event_handler(
     match event {
         FullEvent::ReactionAdd { add_reaction } => {
             let message_id = MessageId::new(ARCHIVE_MESSAGE_ID);
-            if add_reaction.message_id == message_id && data.reaction_roles.contains_key(&add_reaction.emoji) {
-                    if let Some(guild_id) = add_reaction.guild_id {
-                        // TODO: Use try_join to await concurrently?
-                        if let Ok(member) =
-                            guild_id.member(ctx, add_reaction.user_id.unwrap()).await
+            if add_reaction.message_id == message_id
+                && data.reaction_roles.contains_key(&add_reaction.emoji)
+            {
+                if let Some(guild_id) = add_reaction.guild_id {
+                    // TODO: Use try_join to await concurrently?
+                    if let Ok(member) = guild_id.member(ctx, add_reaction.user_id.unwrap()).await {
+                        if let Err(e) = member
+                            .add_role(
+                                &ctx.http,
+                                data.reaction_roles
+                                    .get(&add_reaction.emoji)
+                                    .expect("Hard coded value verified earlier."),
+                            )
+                            .await
                         {
-                            if let Err(e) = member.add_role(&ctx.http, data.reaction_roles.get(&add_reaction.emoji).expect("Hard coded value verified earlier.")).await {
-                                eprintln!("Error: {:?}", e);
-                            }
+                            eprintln!("Error: {:?}", e);
                         }
                     }
                 }
+            }
         }
 
         FullEvent::ReactionRemove { removed_reaction } => {
-                let message_id = MessageId::new(ARCHIVE_MESSAGE_ID);
-                if message_id == removed_reaction.message_id && data.reaction_roles.contains_key(&removed_reaction.emoji) {
-                    if let Some(guild_id) = removed_reaction.guild_id {
-                        if let Ok(member) = guild_id
-                            .member(ctx, removed_reaction.user_id.unwrap())
+            let message_id = MessageId::new(ARCHIVE_MESSAGE_ID);
+            if message_id == removed_reaction.message_id
+                && data.reaction_roles.contains_key(&removed_reaction.emoji)
+            {
+                if let Some(guild_id) = removed_reaction.guild_id {
+                    if let Ok(member) = guild_id
+                        .member(ctx, removed_reaction.user_id.unwrap())
+                        .await
+                    {
+                        if let Err(e) = member
+                            .remove_role(
+                                &ctx.http,
+                                *data
+                                    .reaction_roles
+                                    .get(&removed_reaction.emoji)
+                                    .expect("Hard coded value verified earlier"),
+                            )
                             .await
                         {
-                            if let Err(e) = member.remove_role(&ctx.http, *data.reaction_roles.get(&removed_reaction.emoji).expect("Hard coded value verified earlier")).await {
-                                eprintln!("Error: {:?}", e);
-                            }
+                            eprintln!("Error: {:?}", e);
+                        }
                     }
                 }
             }
