+++ src/scheduler/scheduler.rs
@@ -15,7 +15,7 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
-use crate::tasks::{get_tasks, Task};
+use crate::scheduler::tasks::{get_tasks, Task};
 use serenity::client::Context as SerenityContext;
 
 use tokio::spawn;
