+++ src/scheduler/tasks/tasks.rs
@@ -1,0 +1,52 @@
+/*
+amFOSS Daemon: A discord bot for the amFOSS Discord server.
+Copyright (C) 2024 amFOSS
+
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program.  If not, see <https://www.gnu.org/licenses/>.
+*/
+use crate::utils::time::time_until;
+
+use async_trait::async_trait;
+use serenity::client::Context;
+use tokio::time::Duration;
+
+use super::status_update;
+
+#[async_trait]
+pub trait Task: Send + Sync {
+    fn name(&self) -> &'static str;
+    fn run_in(&self) -> Duration;
+    async fn run(&self, ctx: Context);
+}
+
+pub struct StatusUpdateCheck;
+
+#[async_trait]
+impl Task for StatusUpdateCheck {
+    fn name(&self) -> &'static str {
+        "StatusUpdateCheck"
+    }
+
+    fn run_in(&self) -> Duration {
+        time_until(5, 0)
+    }
+
+    async fn run(&self, ctx: Context) {
+        status_update::check_status_updates(ctx).await;
+    }
+}
+
+pub fn get_tasks() -> Vec<Box<dyn Task>> {
+    vec![Box::new(StatusUpdateCheck)]
+}
